+++ source/js/buyers-guide/bg-main.js
@@ -11,7 +11,7 @@ import JoinUs from "../components/join/join.jsx";
 
 import copyToClipboard from "../../js/copy-to-clipboard.js";
 import HomepageSlider from "./homepage-c-slider.js";
-import ProductGA from "./product-analytics.js";
+import AnalyticsEvents from "./analytics-events.js";
 
 // Track all ReactDOM.render calls so we can use a Promise.all()
 // all the way at the end to make sure we don't report "we are done"
@@ -54,7 +54,7 @@ let main = {
         let gaIdentifier = gaMeta.getAttribute(`content`);
         ReactGA.initialize(gaIdentifier); // UA-87658599-6 by default
         ReactGA.pageview(window.location.pathname);
-        ProductGA.init();
+        AnalyticsEvents.init();
       } else {
         console.warn(`No GA identifier found: skipping bootstrap step`);
       }
